+++ conf/settings.py
@@ -361,21 +361,28 @@ AWS_QUERYSTRING_AUTH = env.bool('AWS_QUERYSTRING_AUTH', False)
 S3_USE_SIGV4 = env.bool('S3_USE_SIGV4', True)
 AWS_S3_HOST = env.str('AWS_S3_HOST', 's3.eu-west-1.amazonaws.com')
 
+AWS_S3_DEFAULT_BINDING_BUCKET_NAME = env.str('AWS_S3_DEFAULT_BINDING_BUCKET_NAME')
+AWS_S3_DATA_SCIENCE_BINDING_BUCKET_NAME = env.str('AWS_S3_DATASCIENCE_BINDING_BUCKET_NAME')
+
 if 'aws-s3-bucket' in VCAP_SERVICES:
-    credentials = VCAP_SERVICES['aws-s3-bucket'][0]['credentials']
-    AWS_ACCESS_KEY_ID = credentials['aws_access_key_id']
-    AWS_SECRET_ACCESS_KEY = credentials['aws_secret_access_key']
-    AWS_STORAGE_BUCKET_NAME = credentials['bucket_name']
-    AWS_S3_REGION_NAME = credentials['aws_region']
-    AWS_S3_ENCRYPTION = True
-    AWS_DEFAULT_ACL = None
-else:
-    AWS_S3_ENCRYPTION = False
-    AWS_ACCESS_KEY_ID = env.str('AWS_ACCESS_KEY_ID', '')
-    AWS_SECRET_ACCESS_KEY = env.str('AWS_SECRET_ACCESS_KEY', '')
-    AWS_STORAGE_BUCKET_NAME = env.str('AWS_STORAGE_BUCKET_NAME', '')
-    AWS_S3_REGION_NAME = env.str('AWS_S3_REGION_NAME', '')
-    AWS_DEFAULT_ACL = 'public-read'
+    for aws_paas_bucket in VCAP_SERVICES['aws-s3-bucket']:
+        if aws_paas_bucket['name'] == AWS_S3_DEFAULT_BINDING_BUCKET_NAME:
+            credentials = aws_paas_bucket['credentials']
+            AWS_ACCESS_KEY_ID = credentials['aws_access_key_id']
+            AWS_SECRET_ACCESS_KEY = credentials['aws_secret_access_key']
+            AWS_STORAGE_BUCKET_NAME = credentials['bucket_name']
+            AWS_S3_REGION_NAME = credentials['aws_region']
+            AWS_S3_ENCRYPTION = True
+            AWS_DEFAULT_ACL = None
+        elif aws_paas_bucket['name'] == AWS_S3_DATA_SCIENCE_BINDING_BUCKET_NAME:
+            credentials = aws_paas_bucket['credentials']
+            AWS_ACCESS_KEY_ID_DATA_SCIENCE = credentials['aws_access_key_id']
+            AWS_SECRET_ACCESS_KEY_DATA_SCIENCE = credentials['aws_secret_access_key']
+            AWS_STORAGE_BUCKET_NAME_DATA_SCIENCE = credentials['bucket_name']
+            AWS_S3_REGION_NAME_DATA_SCIENCE = credentials['aws_region']
+            AWS_S3_ENCRYPTION_DATA_SCIENCE = True
+            AWS_DEFAULT_ACL_DATA_SCIENCE = None
+
 # Admin proxy
 USE_X_FORWARDED_HOST = True
 SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')