+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - TT-1734 - Allow adding second user as member to a company profile
 - TT-1716 - Company search stop words
 - TT-1761 - Remove fab casestudies notification
+ 
+## Breaking changes
+- TT-1538 - AWS-S3 Pass Bucket setup - Pass is required to be setup in target env and set env variable
+    AWS_S3_DEFAULT_BINDING_BUCKET_NAME, AWS_S3_DATASCIENCE_BINDING_BUCKET_NAME
 
 ### Fixed bugs
 - No ticket - Upgrade django to 1.11.23 to fix vulnerability